requirements.txt
numpy
tensorflow
tensorflow-gpu
matplotlib
sklearn
scikit-learn
keras
visualkeras
pillow
opencv-python
imageio
scipy
albumentations